% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplot.R
\name{addseason}
\alias{addseason}
\title{addseson}
\usage{
addseason(data, origin, nodata = TRUE)
}
\arguments{
\item{data}{data frame}

\item{origin}{Origin Date as a string}

\item{nodata}{logical value}
}
\description{
Adding season attribute to a data frame
}

